Add files via upload
The files have not been accordance with naming convention of Java. These file names also not aligned with class name of the program. These are considered as text files with documented code. If you are interested to try them, you can copy the code to make new code files and run. Let me know if there are any faults in it. I will change it.